BaseAnalysis -> AbstractBaseAnalysis
Down to AbstractAnalysis, then to AbstractRoutineAnalysis

AbstractBaseAnalysis -> AbstractAnalysis, AnalysisService
AbstractAnalysis -> AbstractRoutineAnalysis, ReferenceAnalysis
AbstractRoutineAnalysis -> Analysis, DuplicateAnalysis, RejectAnalysis

Lots of little refactoring as I went along.
