+++ bika/lims/browser/analysisrequest/templates/reports/bydepartment.pt
@@ -304,7 +304,7 @@
 <!-- Attachments section -->
 <div id="section-attachments"
     tal:define="ar_attachments python:(analysisrequest['obj'].getAttachment() and len(analysisrequest['obj'].getAttachment())>0) and analysisrequest['obj'].getAttachment() or None;
-                an_attachments python:[an.getAttachment() for an in analysisrequest['obj'].getAnalyses(full_objects=True) if (an.getAttachment() and an.getDepartment() and an.getService().getDepartment().UID()==dept.UID())];">
+                an_attachments python:[an.getAttachment() for an in analysisrequest['obj'].getAnalyses(full_objects=True) if (an.getAttachment() and an.getDepartment() and an.getDepartmentUID()==dept.UID())];">
     <h1 i18n:translate="" tal:condition="python: ar_attachments or len(an_attachments)>0">Attachments</h1>
     <h2 i18n:translate="" tal:condition="ar_attachments">Analysis request attachments</h2>
     <ul tal:condition="ar_attachments">
