+++ bika/lims/browser/reports/selection_macros/select_reference_sample.pt
@@ -5,8 +5,8 @@
     <label i18n:translate="">Reference Sample</label>
     <br/>
 
-    <select name="ReferenceSampleUID:ignore_empty"
-            id="ReferenceSampleUID"
+    <select name="SampleUID:ignore_empty"
+            id="SampleUID"
             tal:attributes="style string:font-family:${here/base_properties/fontFamily};;font-size:100%;">
 
         <option value=""></option>
