+++ bika/lims/content/abstractroutineanalysis.py
@@ -1,0 +1,337 @@
+# -*- coding: utf-8 -*-
+
+# This file is part of Bika LIMS
+#
+# Copyright 2011-2016 by it's authors.
+# Some rights reserved. See LICENSE.txt, AUTHORS.txt.
+
+import cgi
+import math
+from decimal import Decimal
+
+from AccessControl import ClassSecurityInfo
+from DateTime import DateTime
+from Products.Archetypes.Field import StringField, BooleanField, IntegerField, \
+    TextField, FixedPointField
+from Products.Archetypes.Schema import Schema
+from Products.CMFCore.WorkflowCore import WorkflowException
+from bika.lims import bikaMessageFactory as _, deprecated
+from bika.lims import logger
+from bika.lims.browser.fields import UIDReferenceField, DateTimeField
+from bika.lims.browser.widgets import DateTimeWidget, IntegerWidget, \
+    DecimalWidget
+from bika.lims.content.abstractanalysis import AbstractAnalysis
+from bika.lims.content.abstractanalysis import schema
+from bika.lims.content.reflexrule import doReflexRuleAction
+from bika.lims.interfaces import IAnalysis, ISamplePrepWorkflow
+from bika.lims.permissions import Verify as VerifyPermission
+from bika.lims.utils import changeWorkflowState, formatDecimalMark
+from bika.lims.utils import drop_trailing_zeros_decimal
+from bika.lims.utils.analysis import create_analysis, format_numeric_result
+from bika.lims.utils.analysis import get_significant_digits
+from bika.lims.workflow import getTransitionDate, skip
+from plone.api.portal import get_tool
+from plone.api.user import has_permission
+from zope.interface import implements
+
+# Attachments which are added manually in the UI, or automatically when
+# results are imported from a file supplied by an instrument.
+Attachment = UIDReferenceField(
+    'Attachment',
+    multiValued=1,
+    allowed_types=('Attachment',)
+)
+
+# The final result of the analysis is stored here.  The field contains a
+# String value, but the result itself is required to be numeric.  If
+# a non-numeric result is needed, ResultOptions can be used.
+Result = StringField(
+    'Result'
+)
+
+# When the result is changed, this value is updated to the current time.
+# Only the most recent result capture date is recorded here and used to
+# populate catalog values, however the workflow review_history can be
+# used to get all dates of result capture
+ResultCaptureDate = DateTimeField(
+    'ResultCaptureDate'
+)
+
+# If ReportDryMatter is True in the AnalysisService, the adjusted result
+# is stored here.
+ResultDM = StringField(
+    'ResultDM'
+)
+
+# If the analysis has previously been retracted, this flag is set True
+# to indicate that this is a re-test.
+Retested = BooleanField(
+    'Retested',
+    default=False
+)
+
+# When the AR is published, the date of publication is recorded here.
+# It's used to populate catalog values.
+DateAnalysisPublished = DateTimeField(
+    'DateAnalysisPublished',
+    widget=DateTimeWidget(
+        label=_("Date Published")
+    )
+)
+
+# DueDate is calculated by adding the MaxTimeAllowed to the date that the
+# Analysis' sample was received.  It's used to create alerts when analyses
+# are "late", and also to populate catalog values.
+DueDate = DateTimeField(
+    'DueDate',
+    widget=DateTimeWidget(
+        label=_("Due Date")
+    )
+)
+
+# This is used to calculate turnaround time reports.
+# The value is set when the Analysis is published.
+Duration = IntegerField(
+    'Duration',
+    widget=IntegerWidget(
+        label=_("Duration")
+    )
+)
+
+# This is used to calculate turnaround time reports.
+# The value is set when the Analysis is published.
+Earliness = IntegerField(
+    'Earliness',
+    widget=IntegerWidget(
+        label=_("Earliness")
+    )
+)
+
+# The ID of the logged in user who submitted the result for
+# this Analysis.
+Analyst = StringField(
+    'Analyst'
+)
+
+# Remarks entered in the manage_results screen are stored here.  This
+# is also used for remarks generated by results import, or other automated
+# processes
+Remarks = TextField(
+    'Remarks'
+)
+
+# The physical sample partition linked to the Analysis.
+SamplePartition = UIDReferenceField(
+    'SamplePartition',
+    required=0,
+    allowed_types=('SamplePartition',)
+)
+
+# True if the analysis is created by a reflex rule
+IsReflexAnalysis = BooleanField(
+    'IsReflexAnalysis',
+    default=False,
+    required=0
+)
+
+# This field contains the original analysis which was reflected
+OriginalReflexedAnalysis = UIDReferenceField(
+    'OriginalReflexedAnalysis',
+    required=0,
+    allowed_types=('Analysis',)
+)
+
+# This field contains the analysis which has been reflected following
+# a reflex rule
+ReflexAnalysisOf = UIDReferenceField(
+    'ReflexAnalysisOf',
+    required=0,
+    allowed_types=('Analysis',)
+)
+
+# Which is the Reflex Rule action that has created this analysis
+ReflexRuleAction = StringField(
+    'ReflexRuleAction',
+    required=0,
+    default=0
+)
+
+# Which is the 'local_id' inside the reflex rule
+ReflexRuleLocalID = StringField(
+    'ReflexRuleLocalID',
+    required=0,
+    default=0
+)
+
+# Reflex rule triggered actions which the current analysis is responsible for.
+# Separated by '|'
+ReflexRuleActionsTriggered = StringField(
+    'ReflexRuleActionsTriggered',
+    required=0,
+    default=''
+)
+
+# The actual uncertainty for this analysis' result, populated when the result
+# is submitted.
+Uncertainty = FixedPointField(
+    'Uncertainty',
+    precision=10,
+    widget=DecimalWidget(
+        label=_("Uncertainty")
+    )
+)
+
+schema = schema.copy() + Schema((
+    IsReflexAnalysis,
+    OriginalReflexedAnalysis,
+    ReflexAnalysisOf,
+    ReflexRuleAction,
+    ReflexRuleActionsTriggered,
+    ReflexRuleLocalID,
+    SamplePartition,
+    Uncertainty,
+))
+
+
+class AbstractRoutineAnalysis(AbstractAnalysis):
+    implements(IAnalysis, ISamplePrepWorkflow)
+    security = ClassSecurityInfo()
+    displayContentsTab = False
+    schema = schema
+
+    def getSample(self):
+        raise NotImplementedError("getSample is not implemented.")
+
+    @security.public
+    def getSampleUID(self):
+        """Instances must implement getSample
+        """
+        sample = self.getSample()
+        if sample:
+            return sample.UID()
+
+    @security.public
+    def getRequestID(self):
+        """Used to populate catalog values.
+        Returns the ID of the parent analysis request.
+        """
+        ar = self.aq_parent
+        if ar:
+            return ar.getId()
+
+    @security.public
+    def getClientTitle(self):
+        """Used to populate catalog values.
+        Returns the Title of the client for this analysis' AR.
+        """
+        client = self.aq_parent.getClient()
+        if client:
+            return client.Title()
+
+    @security.public
+    def getClientUID(self):
+        """Used to populate catalog values.
+        Returns the UID of the client for this analysis' AR.
+        """
+        client = self.aq_parent.getClient()
+        if client:
+            return client.UID()
+
+    @security.public
+    def getClientURL(self):
+        """This method is used to populate catalog values
+        Returns the URL of the client for this analysis' AR.
+        """
+        client = self.aq_parent.getClient()
+        if client:
+            return client.absolute_url_path()
+
+    @security.public
+    def getClientOrderNumber(self):
+        """Used to populate catalog values.
+        Returns the ClientOrderNumber of the associated AR
+        """
+        return self.aq_parent.getClientOrderNumber()
+
+    @security.public
+    def getDateReceived(self):
+        """Used to populate catalog values.
+        Returns the date on which the "receive" transition was invoked on this
+        analysis' Sample.
+        """
+        sample = self.getSample()
+        if sample:
+            getTransitionDate(self, 'receive')
+
+    @security.public
+    def getDateSampled(self):
+        """Used to populate catalog values.
+        """
+        sample = self.getSample()
+        if sample:
+            getTransitionDate(self, 'sample')
+
+    @security.public
+    def getSamplePointUID(self):
+        """Used to populate catalog values.
+        """
+        sample = self.getSample()
+        if sample:
+            samplepoint = sample.getSamplePoint()
+            if samplepoint:
+                return samplepoint.UID()
+
+    @security.public
+    def getSamplePartitionID(self):
+        """Used to populate catalog values.
+        Returns the sample partition ID
+        """
+        partition = self.getSamplePartition()
+        if partition:
+            return partition.getId()
+        return ''
+
+    @security.public
+    def updateDueDate(self):
+        maxtime = self.getMaxTimeAllowed()
+        if not maxtime:
+            maxtime = {'days': 0, 'hours': 0, 'minutes': 0}
+        max_days = float(maxtime.get('days', 0)) + (
+            (float(maxtime.get('hours', 0)) * 3600 +
+             float(maxtime.get('minutes', 0)) * 60)
+            / 86400
+        )
+        part = self.getSamplePartition()
+        if part:
+            starttime = part.getDateReceived()
+            duetime = starttime + max_days if starttime else ''
+            self.setDueDate(duetime)
+
+    @security.public
+    def getAnalysisRequestTitle(self):
+        """This is a catalog metadata column
+        """
+        return self.aq_parent.Title()
+
+    @security.public
+    def getAnalysisRequestUID(self):
+        """This method is used to populate catalog values
+        """
+        return self.aq_parent.UID()
+
+    @security.public
+    def getAnalysisRequestURL(self):
+        """This is a catalog metadata column
+        """
+        return self.aq_parent.absolute_url_path()
+
+    @security.public
+    def getSampleTypeUID(self):
+        """Used to populate catalog values.
+        """
+        sample = self.getSample()
+        if sample:
+            sampletype = sample.getSampleType()
+            if sampletype:
+                return sampletype.UID()
+
