+++ bika/lims/content/analysisrequest.py
@@ -1870,7 +1870,7 @@ class AnalysisRequest(BaseFolder):
         analyses = self.getAnalyses(full_objects=True)
         value = []
         for analysis in analyses:
-            val = analysis.getServiceTitle()
+            val = analysis.Title()
             if val not in value:
                 value.append(val)
         return value
