+++ bika/lims/content/duplicateanalysis.py
@@ -3,240 +3,70 @@
 # Copyright 2011-2016 by it's authors.
 # Some rights reserved. See LICENSE.txt, AUTHORS.txt.
 
-"""DuplicateAnalysis uses Analysis as it's base.  Until that's fixed there
-is some confusion.
-"""
 from AccessControl import ClassSecurityInfo
-from bika.lims import bikaMessageFactory as _
+from Products.Archetypes.public import Schema, registerType
 from bika.lims.browser.fields import UIDReferenceField
-from bika.lims.utils import t
-from bika.lims.browser.fields import InterimFieldsField
 from bika.lims.config import PROJECTNAME
-from bika.lims.content.analysis import schema, Analysis
+from bika.lims.content.abstractroutineanalysis import AbstractRoutineAnalysis
+from bika.lims.content.abstractroutineanalysis import schema
 from bika.lims.interfaces import IDuplicateAnalysis
 from bika.lims.subscribers import skip
-from Products.Archetypes.config import REFERENCE_CATALOG
-from Products.Archetypes.public import *
-from Products.Archetypes.references import HoldingReference
-from Products.CMFCore.utils import getToolByName
+from plone.api.portal import get_tool
 from zope.interface import implements
 
-AnalysisField = UIDReferenceField(
+# A reference back to the original analysis from which this one was duplicated.
+Analysis = UIDReferenceField(
     'Analysis',
     required=1,
-    allowed_types=('Analysis',),
-)
-
-InterimFields = InterimFieldsField(
-    'InterimFields',
-)
-
-Result = StringField(
-    'Result',
-)
-
-ResultDM = StringField(
-    'ResultDM',
-)
-
-Retested = BooleanField(
-    'Retested',
-)
-
-Analyst = StringField(
-    'Analyst',
-)
-
-SamplePartition = ComputedField(
-    'SamplePartition',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getSamplePartition()',
-)
-
-ClientOrderNumber = ComputedField(
-    'ClientOrderNumber',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getClientOrderNumber()',
-)
-
-Service = ComputedField(
-    'Service',
-    expression='context.getAnalysis() and context.getAnalysis().getService() '
-               'or ""',
-)
-
-ServiceUID = ComputedField(
-    'ServiceUID',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').getServiceUID()',
-)
-
-CategoryUID = ComputedField(
-    'CategoryUID',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getCategoryUID()',
-)
-
-Calculation = ComputedField(
-    'Calculation',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getCalculation()',
-)
-
-ReportDryMatter = ComputedField(
-    'ReportDryMatter',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getReportDryMatter()',
-)
-
-DateReceived = ComputedField(
-    'DateReceived',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getDateReceived()',
-)
-
-MaxTimeAllowed = ComputedField(
-    'MaxTimeAllowed',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getMaxTimeAllowed()',
-)
-
-DueDate = ComputedField(
-    'DueDate',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getDueDate()',
-)
-
-Duration = ComputedField(
-    'Duration',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getDuration()',
-)
-
-Earliness = ComputedField(
-    'Earliness',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getEarliness()',
-)
-
-ClientUID = ComputedField(
-    'ClientUID',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getClientUID()',
-)
-
-RequestID = ComputedField(
-    'RequestID',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').aq_parent.portal_type=="AnalysisRequest" and '
-               'context.getAnalysis().getRequestID() or ""',
-)
-
-PointOfCapture = ComputedField(
-    'PointOfCapture',
-    expression='context.getAnalysis() and context.getAnalysis('
-               ').getPointOfCapture()',
-)
-
-ReferenceAnalysesGroupID = StringField(
-    'ReferenceAnalysesGroupID',
-    widget=StringWidget(
-        label=_("ReferenceAnalysesGroupID"),
-        visible=False,
-    ),
-)
-
-Keyword = ComputedField(
-    'Keyword',
-    expression="context.getAnalysis().getKeyword()",
-)
-
-NumberOfRequiredVerifications = ComputedField(
-    'NumberOfRequiredVerifications',
-    expression='context.getAnalysis().getNumberOfRequiredVerifications()',
+    allowed_types=('Analysis', 'ReferenceAnalysis'),
 )
 
 schema = schema.copy() + Schema((
-    AnalysisField,
-    InterimFields,
-    Result,
-    ResultDM,
-    Retested,
-    Analyst,
-    SamplePartition,
-    ClientOrderNumber,
-    Service,
-    ServiceUID,
-    CategoryUID,
-    Calculation,
-    ReportDryMatter,
-    DateReceived,
-    MaxTimeAllowed,
-    DueDate,
-    Duration,
-    Earliness,
-    ClientUID,
-    RequestID,
-    PointOfCapture,
-    ReferenceAnalysesGroupID,
-    Keyword,
-    NumberOfRequiredVerifications,
+    Analysis,
 ))
 
 
-class DuplicateAnalysis(Analysis):
+class DuplicateAnalysis(AbstractRoutineAnalysis):
     implements(IDuplicateAnalysis)
     security = ClassSecurityInfo()
     displayContentsTab = False
     schema = schema
 
-    _at_rename_after_creation = True
-
-    def _renameAfterCreation(self, check_auto_id=False):
-        from bika.lims.idserver import renameAfterCreation
-        renameAfterCreation(self)
-
-    def getSample(self):
-        analysis = self.getAnalysis()
-        # check for getSample access, eg. ReferenceAnalysis
-        if hasattr(analysis, 'getSample'):
-            return analysis.getSample()
-        # traverse to parent for getSample access
-        analysis_parent = analysis.aq_parent
-        if hasattr(analysis_parent, 'getSample'):
-            return analysis_parent.getSample()
-
+    @security.public
     def getAnalysisPortalType(self):
+        """This returns the portal_type of the original analysis.
         """
-        It is sued as metacolumn
-        """
-        return self.getAnalysis().portal_type
+        analysis = self.getAnalysis()
+        if analysis:
+            return analysis.portal_type
 
+    @security.public
+    def getSample(self):
+        analysis = self.aq_parent
+        return analysis.getSample()
+
+
+
+
+
+    @security.public
     def workflow_script_submit(self):
-        workflow = getToolByName(self, 'portal_workflow')
+        if skip(self, "attach"):
+            return
+        workflow = get_tool('portal_workflow')
         self.reindexObject(idxs=["review_state", ])
         # If all analyses on the worksheet have been submitted,
         # then submit the worksheet.
         ws = self.getBackReferences('WorksheetAnalysis')
         ws = ws[0]
-        # if the worksheet analyst is not assigned, the worksheet can't  be transitioned.
+        # if the worksheet analyst is not assigned, the worksheet can't  be
+        # transitioned.
         if ws.getAnalyst() and not skip(ws, "submit", peek=True):
             all_submitted = True
             for a in ws.getAnalyses():
-                if workflow.getInfoFor(a, 'review_state') in \
-                   ('to_be_sampled', 'to_be_preserved', 'sample_due',
-                    'sample_received', 'assigned',):
+                state = workflow.getInfoFor(a, 'review_state')
+                if state in ('to_be_sampled', 'to_be_preserved', 'sample_due',
+                             'sample_received', 'assigned'):
                     all_submitted = False
                     break
             if all_submitted:
@@ -249,10 +79,11 @@ class DuplicateAnalysis(Analysis):
         if can_attach:
             workflow.doActionFor(self, 'attach')
 
+    @security.public
     def workflow_script_attach(self):
         if skip(self, "attach"):
             return
-        workflow = getToolByName(self, 'portal_workflow')
+        workflow = get_tool('portal_workflow')
         self.reindexObject(idxs=["review_state", ])
         # If all analyses on the worksheet have been attached,
         # then attach the worksheet.
@@ -262,20 +93,19 @@ class DuplicateAnalysis(Analysis):
         if ws_state == 'attachment_due' and not skip(ws, "attach", peek=True):
             can_attach = True
             for a in ws.getAnalyses():
-                if workflow.getInfoFor(a, 'review_state') in \
-                   ('to_be_sampled', 'to_be_preserved', 'sample_due',
-                    'sample_received', 'attachment_due', 'assigned',):
+                state = workflow.getInfoFor(a, 'review_state')
+                if state in ('to_be_sampled', 'to_be_preserved', 'sample_due',
+                             'sample_received', 'attachment_due', 'assigned'):
                     can_attach = False
                     break
             if can_attach:
                 workflow.doActionFor(ws, 'attach')
 
-        return
-
+    @security.public
     def workflow_script_retract(self):
         if skip(self, "retract"):
             return
-        workflow = getToolByName(self, 'portal_workflow')
+        workflow = get_tool('portal_workflow')
         # Escalate action to the Worksheet.
         ws = self.getBackReferences('WorksheetAnalysis')
         ws = ws[0]
@@ -283,15 +113,18 @@ class DuplicateAnalysis(Analysis):
             if workflow.getInfoFor(ws, 'review_state') == 'open':
                 skip(ws, "retract")
             else:
-                if not "retract all analyses" in self.REQUEST['workflow_skiplist']:
-                    self.REQUEST["workflow_skiplist"].append("retract all analyses")
+                if "retract all analyses" \
+                        not in self.REQUEST['workflow_skiplist']:
+                    self.REQUEST["workflow_skiplist"].append(
+                        "retract all analyses")
                 workflow.doActionFor(ws, 'retract')
         self.reindexObject()
 
+    @security.public
     def workflow_script_verify(self):
         if skip(self, "verify"):
             return
-        workflow = getToolByName(self, 'portal_workflow')
+        workflow = get_tool('portal_workflow')
         # If all other analyses on the worksheet are verified,
         # then verify the worksheet.
         ws = self.getBackReferences('WorksheetAnalysis')
@@ -300,22 +133,26 @@ class DuplicateAnalysis(Analysis):
         if ws_state == 'to_be_verified' and not skip(ws, "verify", peek=True):
             all_verified = True
             for a in ws.getAnalyses():
-                if workflow.getInfoFor(a, 'review_state') in \
-                   ('to_be_sampled', 'to_be_preserved', 'sample_due',
-                    'sample_received', 'attachment_due', 'to_be_verified', 'assigned'):
+                state = workflow.getInfoFor(a, 'review_state')
+                if state in ('to_be_sampled', 'to_be_preserved', 'sample_due',
+                             'sample_received', 'attachment_due', 'assigned',
+                             'to_be_verified'):
                     all_verified = False
                     break
             if all_verified:
-                if not "verify all analyses" in self.REQUEST['workflow_skiplist']:
-                    self.REQUEST["workflow_skiplist"].append("verify all analyses")
+                if "verify all analyses" \
+                        not in self.REQUEST['workflow_skiplist']:
+                    self.REQUEST["workflow_skiplist"].append(
+                        "verify all analyses")
                 workflow.doActionFor(ws, "verify")
         self.reindexObject()
 
+    @security.public
     def workflow_script_assign(self):
         if skip(self, "assign"):
             return
-        workflow = getToolByName(self, 'portal_workflow')
-        rc = getToolByName(self, REFERENCE_CATALOG)
+        workflow = get_tool('portal_workflow')
+        rc = get_tool("reference_catalog")
         wsUID = self.REQUEST['context_uid']
         ws = rc.lookupObject(wsUID)
         # retract the worksheet to 'open'
@@ -328,18 +165,20 @@ class DuplicateAnalysis(Analysis):
             workflow.doActionFor(ws, 'retract')
         self.reindexObject()
 
+    @security.public
     def workflow_script_unassign(self):
         if skip(self, "unassign"):
             return
-        workflow = getToolByName(self, 'portal_workflow')
-        rc = getToolByName(self, REFERENCE_CATALOG)
+        workflow = get_tool('portal_workflow')
+        rc = get_tool("reference_catalog")
         wsUID = self.REQUEST['context_uid']
         ws = rc.lookupObject(wsUID)
 
         # May need to promote the Worksheet's review_state
         #  if all other analyses are at a higher state than this one was.
         # (or maybe retract it if there are no analyses left)
-        # Note: duplicates, controls and blanks have 'assigned' as a review_state.
+        # Note: duplicates, controls and blanks have 'assigned' as a
+        # review_state.
         can_submit = True
         can_attach = True
         can_verify = True
@@ -348,41 +187,43 @@ class DuplicateAnalysis(Analysis):
         # We flag this worksheet as empty if there is ONE UNASSIGNED
         # analysis left: worksheet.removeAnalysis() hasn't removed it from
         # the layout yet at this stage.
-        if len(analyses) == 1 \
-           and workflow.getInfoFor(analyses[0], 'review_state') == 'unassigned':
+        state = workflow.getInfoFor(analyses[0], 'review_state')
+        if len(analyses) == 1 and state == 'unassigned':
             ws_empty = True
         for a in analyses:
             ws_empty = False
             a_state = workflow.getInfoFor(a, 'review_state')
-            if a_state in \
-               ('assigned', 'sample_due', 'sample_received',):
+            if a_state in ('assigned', 'sample_due', 'sample_received'):
                 can_submit = False
             else:
                 if not ws.getAnalyst():
                     can_submit = False
-            if a_state in \
-               ('assigned', 'sample_due', 'sample_received', 'attachment_due',):
+            if a_state in ('assigned', 'sample_due', 'sample_received',
+                           'attachment_due'):
                 can_attach = False
-            if a_state in \
-               ('assigned', 'sample_due', 'sample_received', 'attachment_due', 'to_be_verified',):
+            if a_state in ('assigned', 'sample_due', 'sample_received',
+                           'attachment_due', 'to_be_verified'):
                 can_verify = False
+        state = workflow.getInfoFor(ws, 'review_state')
         if not ws_empty:
-        # Note: WS adds itself to the skiplist so we have to take it off again
-        #       to allow multiple promotions (maybe by more than one self).
-            if can_submit and workflow.getInfoFor(ws, 'review_state') == 'open':
+            # Note: WS adds itself to the skiplist so we have to take it off
+            # again to allow multiple promotions (maybe by more than one self).
+            # There-from come the unskips below.
+            if can_submit and state == 'open':
                 workflow.doActionFor(ws, 'submit')
                 skip(ws, 'submit', unskip=True)
-            if can_attach and workflow.getInfoFor(ws, 'review_state') == 'attachment_due':
+            if can_attach and state == 'attachment_due':
                 workflow.doActionFor(ws, 'attach')
                 skip(ws, 'attach', unskip=True)
-            if can_verify and workflow.getInfoFor(ws, 'review_state') == 'to_be_verified':
+            if can_verify and state == 'to_be_verified':
                 self.REQUEST["workflow_skiplist"].append('verify all analyses')
                 workflow.doActionFor(ws, 'verify')
                 skip(ws, 'verify', unskip=True)
         else:
-            if workflow.getInfoFor(ws, 'review_state') != 'open':
+            if state != 'open':
                 workflow.doActionFor(ws, 'retract')
                 skip(ws, 'retract', unskip=True)
         self.reindexObject()
+
 
 registerType(DuplicateAnalysis, PROJECTNAME)
