+++ bika/lims/content/referenceanalysis.py
@@ -4,194 +4,82 @@
 #
 # Copyright 2011-2016 by it's authors.
 # Some rights reserved. See LICENSE.txt, AUTHORS.txt.
-
-
-"""ReferenceAnalysis
-"""
-from plone import api
 from AccessControl import ClassSecurityInfo
-from bika.lims import bikaMessageFactory as _
-from bika.lims.utils import t, formatDecimalMark
-from bika.lims.utils.analysis import format_numeric_result
-from bika.lims.browser.fields import HistoryAwareReferenceField
-from bika.lims.browser.fields import InterimFieldsField
-from bika.lims.browser.widgets import RecordsWidget as BikaRecordsWidget
-from bika.lims.config import STD_TYPES, PROJECTNAME
-from bika.lims.content.bikaschema import BikaSchema
+from DateTime import DateTime
+from Products.Archetypes.config import REFERENCE_CATALOG
+from Products.Archetypes.public import *
+from Products.CMFCore.WorkflowCore import WorkflowException
+from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.utils import safe_unicode
+from bika.lims import logger, deprecated
+from bika.lims.browser.fields import UIDReferenceField
+from bika.lims.config import PROJECTNAME, STD_TYPES
+from bika.lims.content.abstractanalysis import AbstractAnalysis
+from bika.lims.content.abstractanalysis import schema
 from bika.lims.interfaces import IReferenceAnalysis
 from bika.lims.permissions import Verify as VerifyPermission
 from bika.lims.subscribers import skip
 from bika.lims.utils.analysis import get_significant_digits
-from DateTime import DateTime
 from plone.app.blob.field import BlobField
-from Products.Archetypes.config import REFERENCE_CATALOG
-from Products.Archetypes.public import *
-from Products.Archetypes.references import HoldingReference
-from Products.ATExtensions.ateapi import DateTimeField
-from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.utils import safe_unicode
 from zope.interface import implements
+from plone.api.user import has_permission
 
-
-schema = BikaSchema.copy() + Schema((
-    StringField('ReferenceType',
+schema = schema.copy() + Schema((
+    StringField(
+        'ReferenceType',
         vocabulary=STD_TYPES,
-        widget=SelectionWidget(
-            format='select',
-            label=_("Reference Type"),
-        ),
     ),
-    HistoryAwareReferenceField('Service',
-        required=1,
-        allowed_types=('AnalysisService',),
-        relationship='ReferenceAnalysisAnalysisService',
-        referenceClass=HoldingReference,
-        widget=ReferenceWidget(
-            label=_("Analysis Service"),
-        )
+    BlobField(
+        'RetractedAnalysesPdfReport',
     ),
-    InterimFieldsField('InterimFields',
-        widget=BikaRecordsWidget(
-            label=_("Calculation Interim Fields"),
-        )
-    ),
-    StringField('Result',
-        widget=StringWidget(
-            label=_("Result"),
-        )
-    ),
-    DateTimeField('ResultCaptureDate',
-        widget=ComputedWidget(
-            visible=False,
-        ),
-    ),
-    StringField('ResultDM',
-    ),
-    ReferenceField('Attachment',
-        multiValued=1,
-        allowed_types=('Attachment',),
-        referenceClass=HoldingReference,
-        relationship='ReferenceAnalysisAttachment',
-    ),
-    StringField('Analyst',
-    ),
-    TextField('Remarks',
-    ),
-    ReferenceField(
-        'Instrument',
-        required=0,
-        allowed_types=('Instrument',),
-        relationship='AnalysisInstrument',
-        referenceClass=HoldingReference,
-    ),
-    ReferenceField('Method',
-        required = 0,
-        allowed_types = ('Method',),
-        relationship = 'AnalysisMethod',
-        referenceClass = HoldingReference,
-    ),
-    BlobField('RetractedAnalysesPdfReport',
-    ),
-    BooleanField('Retested',
-        default = False,
-        widget = BooleanWidget(
-            label=_("Retested"),
-        ),
-    ),
-    ComputedField('ReferenceSampleUID',
-        expression = 'context.aq_parent.UID()',
-        widget = ComputedWidget(
-            visible=False,
-        ),
-    ),
-    ComputedField('SupplierUID',
-        expression = 'context.aq_parent.aq_parent.UID()',
-        widget = ComputedWidget(
-            visible=False,
-        ),
-    ),
-    ComputedField('ServiceUID',
-        expression = "context.getService() and context.getService().UID() or ''",
-        widget = ComputedWidget(
-            visible=False,
-        ),
-    ),
-    StringField('ReferenceAnalysesGroupID',
-        widget = StringWidget(
-            label=_("ReferenceAnalysesGroupID"),
-            visible=False,
-        ),
-    ),
-    ComputedField('Keyword',
-        expression = "context.getService() and context.getService().getKeyword() or ''",
-    ),
-    FixedPointField('Uncertainty',
-        widget=DecimalWidget(
-            label = _("Uncertainty"),
-        ),
-    ),
-
-    # Required number of required verifications before this analysis being
-    # transitioned to a 'verified' state. This value is set automatically
-    # when the analysis is created, based on the value set for the property
-    # NumberOfRequiredVerifications from the Analysis Service
-    IntegerField('NumberOfRequiredVerifications', default=1),
-
-    # This field keeps the user_ids of members who verified this analysis.
-    # After each verification, user_id will be added end of this string
-    # seperated by comma- ',' .
-    StringField('Verificators',default='')
-),
-)
+    StringField(
+        'ReferenceAnalysesGroupID',
+    )
+))
 
 
-class ReferenceAnalysis(BaseContent):
+class ReferenceAnalysis(AbstractAnalysis):
     implements(IReferenceAnalysis)
     security = ClassSecurityInfo()
     displayContentsTab = False
     schema = schema
 
-    _at_rename_after_creation = True
-
-    def _renameAfterCreation(self, check_auto_id=False):
-        from bika.lims.idserver import renameAfterCreation
-        renameAfterCreation(self)
+    def getSupplierUID(self):
+        return self.aq_parent.aq_parent.UID()
 
     def getNumberOfVerifications(self):
-        verificators=self.getVerificators()
+        verificators = self.getVerificators()
         if not verificators:
             return 0
         return len(verificators.split(','))
 
-    def addVerificator(self,username):
-        verificators=self.getVerificators()
+    def addVerificator(self, username):
+        verificators = self.getVerificators()
         if not verificators:
             self.setVerificators(username)
         else:
-            self.setVerificators(verificators+","+username)
+            self.setVerificators(verificators + "," + username)
 
     def deleteLastVerificator(self):
-        verificators=self.getVerificators().split(',')
+        verificators = self.getVerificators().split(',')
         del verificators[-1]
         self.setVerificators(",".join(verificators))
 
-    def wasVerifiedByUser(self,username):
-        verificators=self.getVerificators().split(',')
+    def wasVerifiedByUser(self, username):
+        verificators = self.getVerificators().split(',')
         return username in verificators
 
     def getLastVerificator(self):
         return self.getVerificators().split(',')[-1]
 
+    @deprecated("You should use the Reference Analysis title directly.")
     def Title(self):
-        """ Return the Service ID as title """
+        """Return the Service ID as title
+        Titles of analyses are always the same as the title of the service.
+        """
         s = self.getService()
         s = s and s.Title() or ''
         return safe_unicode(s).encode('utf-8')
-
-    def getReviewState(self):
-        """ Return the current referencenalysis' state"""
-        workflow = getToolByName(self, "portal_workflow")
-        return workflow.getInfoFor(self, "review_state")
 
     def getWorksheetUID(self):
         """
@@ -212,7 +100,8 @@ class ReferenceAnalysis(BaseContent):
         """ Calls self.Service.getUncertainty with either the provided
             result value or self.Result
         """
-        return self.getService().getUncertainty(result and result or self.getResult())
+        return self.getService().getUncertainty(
+            result and result or self.getResult())
 
     def getUncertainty(self, result=None):
         """ Returns the uncertainty for this analysis and result.
@@ -222,7 +111,7 @@ class ReferenceAnalysis(BaseContent):
         """
         serv = self.getService()
         schu = self.Schema().getField('Uncertainty').get(self)
-        if schu and serv.getAllowManualUncertainty() == True:
+        if schu and serv.getAllowManualUncertainty():
             try:
                 schu = float(schu)
                 return schu
@@ -231,7 +120,7 @@ class ReferenceAnalysis(BaseContent):
                 return self.getDefaultUncertainty(result)
         return self.getDefaultUncertainty(result)
 
-    def getSample(self, result=None):
+    def getSample(self):
         """ Conform to Analysis
         """
         return self.aq_parent
@@ -245,7 +134,8 @@ class ReferenceAnalysis(BaseContent):
 
     security.declarePublic('current_date')
 
-    def current_date(self):
+    @staticmethod
+    def current_date():
         """ return current date """
         return DateTime()
 
@@ -258,14 +148,14 @@ class ReferenceAnalysis(BaseContent):
         # TODO : Remove when analysis - instrument being assigned directly
         if not self.getInstrument():
             instr = self.getService().getInstrument() \
-                    if self.getService().getInstrumentEntryOfResults() \
-                    else None
+                if self.getService().getInstrumentEntryOfResults() \
+                else None
             if instr:
                 self.setInstrument(instr)
         # ---8<--------
 
         return self.getInstrument().isValid() \
-                if self.getInstrument() else True
+            if self.getInstrument() else True
 
     def isInstrumentAllowed(self, instrument):
         """ Checks if the specified instrument can be set for this
@@ -289,7 +179,6 @@ class ReferenceAnalysis(BaseContent):
             uid = instrument.UID()
 
         method = self.getMethod()
-        instruments = []
         if not method:
             # Look for Analysis Service methods and instrument support
             instruments = service.getRawInstruments()
@@ -312,7 +201,7 @@ class ReferenceAnalysis(BaseContent):
 
         service = self.getService()
         if service.getManualEntryOfResults() is True \
-            and uid in service.getRawMethods():
+                and uid in service.getRawMethods():
             return True
 
         if service.getInstrumentEntryOfResults() is True:
@@ -322,142 +211,9 @@ class ReferenceAnalysis(BaseContent):
 
         return False
 
-    def getAnalyst(self):
-        """ Returns the identifier of the assigned analyst. If there is
-            no analyst assigned, and this analysis is attached to a
-            worksheet, retrieves the analyst assigned to the parent
-            worksheet
-        """
-        field = self.getField('Analyst')
-        analyst = field and field.get(self) or ''
-        if not analyst:
-            # Is assigned to a worksheet?
-            wss = self.getBackReferences('WorksheetAnalysis')
-            if len(wss) > 0:
-                analyst = wss[0].getAnalyst()
-                field.set(self, analyst)
-        return analyst if analyst else ''
-
-    def getAnalystName(self):
-        """ Returns the name of the currently assigned analyst
-        """
-        mtool = getToolByName(self, 'portal_membership')
-        analyst = self.getAnalyst().strip()
-        analyst_member = mtool.getMemberById(analyst)
-        if analyst_member is not None:
-            return analyst_member.getProperty('fullname')
-        else:
-            return ''
-
-    def getDepartmentUID(self):
-        return self.getService().getDepartment().UID()
-
-    # TODO-performance: improve this function using another catalog and takeing
-    # advantatge of the column in service, not getting the full object.
-    def getCategoryTitle(self):
-        """
-        Returns the Title of the asociated service's department.
-        """
-        # Getting the service like that because otherwise gives an error
-        # when rebuilding the catalogs.
-        service_uid = self.getRawService()
-        catalog = getToolByName(self, "uid_catalog")
-        brain = catalog(UID=service_uid)
-        if brain:
-            return brain[0].getObject().getCategoryTitle()
-        return ''
-
-    def getFormattedResult(self, specs=None, decimalmark='.', sciformat=1):
-        """Formatted result:
-        1. If the result is not floatable, return it without being formatted
-        2. If the analysis specs has hidemin or hidemax enabled and the
-           result is out of range, render result as '<min' or '>max'
-        3. If the result is floatable, render it to the correct precision
-        specs param is optional. A dictionary as follows:
-            {'min': <min_val>,
-             'max': <max_val>,
-             'error': <error>,
-             'hidemin': <hidemin_val>,
-             'hidemax': <hidemax_val>}
-        :param sciformat: 1. The sci notation has to be formatted as aE^+b
-                          2. The sci notation has to be formatted as a·10^b
-                          3. As 2, but with super html entity for exp
-                          4. The sci notation has to be formatted as a·10^b
-                          5. As 4, but with super html entity for exp
-                          By default 1
-        """
-        result = self.getResult()
-        service = self.getService()
-
-        # 1. If the result is not floatable, return it without being formatted
-        try:
-            result = float(result)
-        except:
-            return result
-
-        # 2. If the analysis specs has enabled hidemin or hidemax and the
-        #    result is out of range, render result as '<min' or '>max'
-        belowmin = False
-        abovemax = False
-        if not specs:
-            uid = self.getServiceUID()
-            specs = self.aq_parent.getResultsRangeDict().get(uid, {})
-        hidemin = specs.get('hidemin', '')
-        hidemax = specs.get('hidemax', '')
-        try:
-            belowmin = hidemin and result < float(hidemin) or False
-        except:
-            belowmin = False
-            pass
-        try:
-            abovemax = hidemax and result > float(hidemax) or False
-        except:
-            abovemax = False
-            pass
-
-        # 2.1. If result is below min and hidemin enabled, return '<min'
-        if belowmin:
-            return formatDecimalMark('< %s' % hidemin, decimalmark)
-
-        # 2.2. If result is above max and hidemax enabled, return '>max'
-        if abovemax:
-            return formatDecimalMark('> %s' % hidemax, decimalmark)
-
-        # 3. If the result is floatable, render it to the correct precision
-        return formatDecimalMark(format_numeric_result(self, result, sciformat), decimalmark)
-
-    def getPrecision(self, result=None):
-        """
-        Returns the precision for the Analysis.
-        - ManualUncertainty not set: returns the precision from the
-            AnalysisService.
-        - ManualUncertainty set and Calculate Precision from Uncertainty
-          is also set in Analysis Service: calculates the precision of the
-          result according to the manual uncertainty set.
-        - ManualUncertainty set and Calculatet Precision from Uncertainty
-          not set in Analysis Service: returns the result as-is.
-        Further information at AnalysisService.getPrecision()
-        """
-        serv = self.getService()
-        schu = self.Schema().getField('Uncertainty').get(self)
-        if schu and serv.getAllowManualUncertainty() == True \
-            and serv.getPrecisionFromUncertainty() == True:
-            uncertainty = self.getUncertainty(result)
-            if uncertainty == 0:
-                return 1
-            return get_significant_digits(uncertainty)
-        else:
-            return serv.getPrecision(result)
-
-    def getParentUID(self):
-        """
-        It is used as metacolumn
-        """
-        return self.aq_parent.UID()
 
     def getExpiryDate(self):
-        """
-        It is used as a metacolumn.
+        """It is used as a metacolumn.
         Returns the expiration date from the reference sample.
         """
         return self.aq_parent.getExpiryDate()
@@ -502,7 +258,7 @@ class ReferenceAnalysis(BaseContent):
         """
         # Check if the user has "Bika: Verify" privileges
         username = member.getUserName()
-        allowed = api.user.has_permission(VerifyPermission, username=username)
+        allowed = has_permission(VerifyPermission, username=username)
         if not allowed:
             return False
 
@@ -530,19 +286,25 @@ class ReferenceAnalysis(BaseContent):
         if self_submitted and not selfverification:
             return False
 
-        #Checking verifiability depending on multi-verification type of bika_setup
-        if self.bika_setup.getNumberOfRequiredVerifications()>1:
-            mv_type=self.bika_setup.getTypeOfmultiVerification()
-            #If user verified before and self_multi_disabled, then return False
-            if mv_type=='self_multi_disabled' and self.wasVerifiedByUser(username):
+        # Checking verifiability depending on multi-verification type of
+        # bika_setup
+        if self.bika_setup.getNumberOfRequiredVerifications() > 1:
+            mv_type = self.bika_setup.getTypeOfmultiVerification()
+            # If user verified before and self_multi_disabled, then return False
+            if mv_type == 'self_multi_disabled' and self.wasVerifiedByUser(
+                    username):
                 return False
 
-            # If user is the last verificator and consecutively multi-verification
+            # If user is the last verificator and consecutively
+            # multi-verification
             # is disabled, then return False
-            # Comparing was added just to check if this method is called before/after
+            # Comparing was added just to check if this method is called
+            # before/after
             # verification
-            elif mv_type=='self_multi_not_cons' and username==self.getLastVerificator() and \
-                self.getNumberOfVerifications()<self.getNumberOfRequiredVerifications():
+            elif mv_type == 'self_multi_not_cons' and username == \
+                    self.getLastVerificator() and \
+                            self.getNumberOfVerifications() < \
+                            self.getNumberOfRequiredVerifications():
                 return False
 
         # All checks pass
@@ -556,7 +318,6 @@ class ReferenceAnalysis(BaseContent):
         :returns: true or false
         """
         mtool = getToolByName(self, "portal_membership")
-        checkPermission = mtool.checkPermission
         # Check if the analysis is in a "verifiable" state
         if self.isVerifiable():
             # Check if the user can verify the analysis
@@ -572,12 +333,13 @@ class ReferenceAnalysis(BaseContent):
         # then submit the worksheet.
         ws = self.getBackReferences('WorksheetAnalysis')
         ws = ws[0]
-        # if the worksheet analyst is not assigned, the worksheet can't  be transitioned.
+        # if the worksheet analyst is not assigned, the worksheet can't  be
+        # transitioned.
         if ws.getAnalyst() and not skip(ws, "submit", peek=True):
             all_submitted = True
             for a in ws.getAnalyses():
                 if workflow.getInfoFor(a, 'review_state') in \
-                   ('sample_due', 'sample_received', 'assigned',):
+                        ('sample_due', 'sample_received', 'assigned',):
                     all_submitted = False
                     break
             if all_submitted:
@@ -605,7 +367,8 @@ class ReferenceAnalysis(BaseContent):
             can_attach = True
             for a in ws.getAnalyses():
                 if workflow.getInfoFor(a, 'review_state') in \
-                   ('sample_due', 'sample_received', 'attachment_due', 'assigned',):
+                        ('sample_due', 'sample_received', 'attachment_due',
+                         'assigned',):
                     can_attach = False
                     break
             if can_attach:
@@ -623,8 +386,10 @@ class ReferenceAnalysis(BaseContent):
             if workflow.getInfoFor(ws, 'review_state') == 'open':
                 skip(ws, "retract")
             else:
-                if not "retract all analyses" in self.REQUEST['workflow_skiplist']:
-                    self.REQUEST["workflow_skiplist"].append("retract all analyses")
+                if "retract all analyses" \
+                        not in self.REQUEST['workflow_skiplist']:
+                    self.REQUEST["workflow_skiplist"].append(
+                        "retract all analyses")
                 workflow.doActionFor(ws, 'retract')
         self.reindexObject()
 
@@ -638,16 +403,20 @@ class ReferenceAnalysis(BaseContent):
         if ws and len(ws) > 0:
             ws = ws[0]
             ws_state = workflow.getInfoFor(ws, 'review_state')
-            if ws_state == 'to_be_verified' and not skip(ws, "verify", peek=True):
+            if ws_state == 'to_be_verified' and not skip(ws, "verify",
+                                                         peek=True):
                 all_verified = True
                 for a in ws.getAnalyses():
                     if workflow.getInfoFor(a, 'review_state') in \
-                       ('sample_due', 'sample_received', 'attachment_due', 'to_be_verified', 'assigned'):
+                            ('sample_due', 'sample_received', 'attachment_due',
+                             'to_be_verified', 'assigned'):
                         all_verified = False
                         break
                 if all_verified:
-                    if not "verify all analyses" in self.REQUEST['workflow_skiplist']:
-                        self.REQUEST["workflow_skiplist"].append("verify all analyses")
+                    if "verify all analyses" \
+                            not in self.REQUEST['workflow_skiplist']:
+                        self.REQUEST["workflow_skiplist"].append(
+                            "verify all analyses")
                     workflow.doActionFor(ws, "verify")
         self.reindexObject()
 
@@ -664,9 +433,11 @@ class ReferenceAnalysis(BaseContent):
             ws_state = workflow.getInfoFor(ws, 'review_state')
             if ws_state != 'open':
                 if 'workflow_skiplist' not in self.REQUEST:
-                    self.REQUEST['workflow_skiplist'] = ['retract all analyses', ]
+                    self.REQUEST['workflow_skiplist'] = [
+                        'retract all analyses', ]
                 else:
-                    self.REQUEST["workflow_skiplist"].append('retract all analyses')
+                    self.REQUEST["workflow_skiplist"].append(
+                        'retract all analyses')
                 workflow.doActionFor(ws, 'retract')
         self.reindexObject()
 
@@ -681,7 +452,8 @@ class ReferenceAnalysis(BaseContent):
         # May need to promote the Worksheet's review_state
         #  if all other analyses are at a higher state than this one was.
         # (or maybe retract it if there are no analyses left)
-        # Note: duplicates, controls and blanks have 'assigned' as a review_state.
+        # Note: duplicates, controls and blanks have 'assigned' as a
+        # review_state.
         can_submit = True
         can_attach = True
         can_verify = True
@@ -693,34 +465,42 @@ class ReferenceAnalysis(BaseContent):
         # analysis left: worksheet.removeAnalysis() hasn't removed it from
         # the layout yet at this stage.
         if len(analyses) == 1 \
-           and workflow.getInfoFor(analyses[0], 'review_state') == 'unassigned':
+                and workflow.getInfoFor(analyses[0],
+                                        'review_state') == 'unassigned':
             ws_empty = True
 
         for a in analyses:
             a_state = workflow.getInfoFor(a, 'review_state')
             if a_state in \
-               ('assigned', 'sample_due', 'sample_received',):
+                    ('assigned', 'sample_due', 'sample_received',):
                 can_submit = False
             else:
                 if not ws.getAnalyst():
                     can_submit = False
             if a_state in \
-               ('assigned', 'sample_due', 'sample_received', 'attachment_due',):
+                    ('assigned', 'sample_due', 'sample_received',
+                     'attachment_due',):
                 can_attach = False
             if a_state in \
-               ('assigned', 'sample_due', 'sample_received', 'attachment_due', 'to_be_verified',):
+                    ('assigned', 'sample_due', 'sample_received',
+                     'attachment_due', 'to_be_verified',):
                 can_verify = False
 
         if not ws_empty:
-        # Note: WS adds itself to the skiplist so we have to take it off again
-        #       to allow multiple promotions (maybe by more than one self).
+            # Note: WS adds itself to the skiplist so we have to take it off
+            # again
+            #       to allow multiple promotions (maybe by more than one self).
             if can_submit and workflow.getInfoFor(ws, 'review_state') == 'open':
                 workflow.doActionFor(ws, 'submit')
                 skip(ws, 'submit', unskip=True)
-            if can_attach and workflow.getInfoFor(ws, 'review_state') == 'attachment_due':
+            if can_attach and workflow.getInfoFor(ws,
+                                                  'review_state') == \
+                    'attachment_due':
                 workflow.doActionFor(ws, 'attach')
                 skip(ws, 'attach', unskip=True)
-            if can_verify and workflow.getInfoFor(ws, 'review_state') == 'to_be_verified':
+            if can_verify and workflow.getInfoFor(ws,
+                                                  'review_state') == \
+                    'to_be_verified':
                 self.REQUEST["workflow_skiplist"].append('verify all analyses')
                 workflow.doActionFor(ws, 'verify')
                 skip(ws, 'verify', unskip=True)
