+++ bika/lims/setuphandlers.py
@@ -570,7 +570,6 @@ class BikaGenerator:
         addIndex(bc, 'getSampleTypeUID', 'FieldIndex')
         addIndex(bc, 'getSampleUID', 'FieldIndex')
         addIndex(bc, 'getSamplingDate', 'DateIndex')
-        addIndex(bc, 'getServiceTitle', 'FieldIndex')
         addIndex(bc, 'getWorksheetTemplateTitle', 'FieldIndex')
         addIndex(bc, 'Priority', 'FieldIndex')
         addIndex(bc, 'BatchUID', 'FieldIndex')
@@ -708,7 +707,6 @@ class BikaGenerator:
         addIndex(bsc, 'getSamplePointUID', 'FieldIndex')
         addIndex(bsc, 'getSampleTypeTitle', 'KeywordIndex')
         addIndex(bsc, 'getSampleTypeUID', 'FieldIndex')
-        addIndex(bsc, 'getServiceTitle', 'FieldIndex')
         addIndex(bsc, 'getServiceUID', 'FieldIndex')
         addIndex(bsc, 'getTotalPrice', 'FieldIndex')
         addIndex(bsc, 'getUnit', 'FieldIndex')
@@ -768,7 +766,6 @@ class BikaGenerator:
         addColumn(bsc, 'getSamplePointUID')
         addColumn(bsc, 'getSampleTypeTitle')
         addColumn(bsc, 'getSampleTypeUID')
-        addColumn(bsc, 'getServiceTitle')
         addColumn(bsc, 'getServiceUID')
         addColumn(bsc, 'getTotalPrice')
         addColumn(bsc, 'getUnit')
