+++ bika/lims/utils/analysis.py
@@ -57,7 +57,9 @@ def create_analysis(context, source, **kwargs):
     IGNORE_FIELDNAMES = [
         'UID', 'id', 'allowDiscussion', 'subject', 'location', 'contributors',
         'creators', 'effectiveDate', 'expirationDate', 'language', 'rights',
-        'creation_date', 'modification_date']
+        'creation_date', 'modification_date', 'IsReflexAnalysis',
+        'OriginalReflexedAnalysis', 'ReflexAnalysisOf', 'ReflexRuleAction',
+        'ReflexRuleLocalID', 'ReflexRuleActionsTriggered']
     for field in src_schema.fields():
         fieldname = field.getName()
         if fieldname in IGNORE_FIELDNAMES and fieldname not in kwargs:
